feat(source-generator,EF): Add filter generator to EFCore
This source generator allows that a call to `.ToPredicate()` create a predicate
based on properties values and annotations.

This will avoid that a dev write lots of boiler plate - with `if`s and other decision statements -
just to query a bunch of records.

All generated code does not use reflection, enabling compiler optimizations.

For more details about how to use, check README.md and samples/EfSamples project.